requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
boto3==1.4.4
botocore==1.5.45
cffi==1.10.0
click==6.7
configparser==3.5.0
cryptography==1.8.1
dask==0.14.1
dask-ec2==0.5.0
docutils==0.13.1
enum34==1.1.6
futures==3.1.1
idna==2.5
ipaddress==1.0.18
jmespath==0.9.2
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
voluptuous==0.10.5
